+++ Helper/Config.php
@@ -4,7 +4,6 @@ namespace Dintero\Checkout\Helper;
 
 use Dintero\Checkout\Model\Api\Client;
 use Dintero\Checkout\Model\Dintero;
-use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Helper\Context;
 use Magento\Framework\Encryption\EncryptorInterface;
@@ -86,6 +85,16 @@ class Config extends AbstractHelper
     const XPATH_IS_EMBEDDED = 'payment/dintero/is_embedded';
 
     /*
+     * Embed Type
+     */
+    const XPATH_EMBED_TYPE = 'payment/dintero/embed_type';
+
+    /*
+     * Embed Type
+     */
+    const XPATH_IS_POPOUT = 'payment/dintero/is_popout';
+
+    /*
      * XPATH Express checkout enabled
      */
     const XPATH_IS_EXPRESS = 'payment/dintero/is_express';
@@ -111,6 +120,11 @@ class Config extends AbstractHelper
     const XPATH_UNSPECIFIED_METHODS = 'payment/dintero/unspecified_methods_map';
 
     /*
+     * Allow ship-to-different address
+     */
+    const XPATH_ALLOW_DIFF_SHIP_ADDR = 'payment/dintero/allow_different_shipping';
+
+    /*
      * Default callback delay in seconds
      */
     const DEFAULT_CALLBACK_DELAY = 30;
@@ -129,6 +143,11 @@ class Config extends AbstractHelper
      * Enable pay button on product page
      */
     const XPATH_PRODUCT_PAGE_BUTTON_ENABLED = 'payment/dintero/product_page_button_enabled';
+
+    /*
+     * Enable pay button on minicart and cart
+     */
+    const XPATH_CART_BUTTON_ENABLED = 'payment/dintero/cart_button_enabled';
 
     /*
      * Number of days for session expiration
@@ -162,8 +181,8 @@ class Config extends AbstractHelper
      * @param EncryptorInterface $encryptor
      */
     public function __construct(
-        Context $context,
-        EncryptorInterface $encryptor,
+        Context               $context,
+        EncryptorInterface    $encryptor,
         StoreManagerInterface $storeManager
     ) {
         parent::__construct($context);
@@ -505,6 +524,18 @@ class Config extends AbstractHelper
     }
 
     /**
+     * Check if cart and minicart buttons should be enabled
+     *
+     * @return bool
+     */
+    public function showExpressCartButton()
+    {
+        return $this->isActive()
+            && $this->isExpress()
+            && $this->scopeConfig->isSetFlag(self::XPATH_CART_BUTTON_ENABLED, ScopeInterface::SCOPE_STORE);
+    }
+
+    /**
      * Retrieving payment action
      *
      * @param int|string $scopeCode
@@ -582,7 +613,7 @@ class Config extends AbstractHelper
      */
     public function getSessionExpirationDays()
     {
-        return  $this->scopeConfig->getValue(self::XPATH_SESSION_EXP_DAY);
+        return $this->scopeConfig->getValue(self::XPATH_SESSION_EXP_DAY);
     }
 
     /**
@@ -614,7 +645,7 @@ class Config extends AbstractHelper
     {
         return explode(
             ',',
-            $this->scopeConfig->getValue(self::XPATH_PICKUP_METHODS, ScopeInterface::SCOPE_STORE, $scopeCode)
+            $this->scopeConfig->getValue(self::XPATH_PICKUP_METHODS, ScopeInterface::SCOPE_STORE, $scopeCode) ?? ''
         );
     }
 
@@ -628,7 +659,7 @@ class Config extends AbstractHelper
     {
         return explode(
             ',',
-            $this->scopeConfig->getValue(self::XPATH_UNSPECIFIED_METHODS, ScopeInterface::SCOPE_STORE, $scopeCode)
+            $this->scopeConfig->getValue(self::XPATH_UNSPECIFIED_METHODS, ScopeInterface::SCOPE_STORE, $scopeCode) ?? ''
         );
     }
 
@@ -640,5 +671,51 @@ class Config extends AbstractHelper
     public function getLineIdFieldName()
     {
         return $this->scopeConfig->getValue(self::XPATH_ID_FIELD);
+    }
+
+    /**
+     * Retrieve embed type
+     *
+     * @param string $scopeCode
+     * @return string
+     */
+    public function getEmbedType($scopeCode = null)
+    {
+        $embedType = $this->scopeConfig->getValue(self::XPATH_EMBED_TYPE, ScopeInterface::SCOPE_STORE, $scopeCode);
+        return $embedType === Client::TYPE_EXPRESS ? Client::TYPE_EXPRESS : Client::TYPE_EMBEDDED;
+    }
+
+    /**
+     * Check if embedded express is enabled
+     *
+     * @param string $scopeCode
+     * @return bool
+     */
+    public function isEmbeddedExpress($scopeCode = null)
+    {
+        return $this->getEmbedType($scopeCode) === Client::TYPE_EXPRESS;
+    }
+
+    /**
+     * Check if embedded checkout popout mode is enabled
+     *
+     * @param $scopeCode
+     * @return bool
+     */
+    public function isPopOut($scopeCode = null)
+    {
+        return $this->scopeConfig->isSetFlag(self::XPATH_IS_POPOUT, ScopeInterface::SCOPE_STORE, $scopeCode);
+    }
+
+    /**
+     * Allow ship-to-different address
+     *
+     * @param string $scopeCode
+     * @return string[]
+     */
+    public function getDifferentShippingAddressCustomerTypes($scopeCode = null)
+    {
+        $value = $this->scopeConfig->getValue(self::XPATH_ALLOW_DIFF_SHIP_ADDR, ScopeInterface::SCOPE_STORE, $scopeCode);
+        return $value ? explode(',', $value) : [];
     }
 }