+++ view/frontend/requirejs-config.js
@@ -13,7 +13,7 @@ var config = {
         '*': {
             dinteroProductAdd: 'Dintero_Checkout/js/dintero-product-add',
             dinteroLogo: 'Dintero_Checkout/js/view/logo',
-            dinteroSdk: 'https://unpkg.com/@dintero/checkout-web-sdk@0.0.15/dist/checkout-web-sdk.umd.js'
+            dinteroSdk: 'https://unpkg.com/@dintero/checkout-web-sdk@0.8.0/dist/dintero-checkout-web-sdk.umd.min.js'
         }
     }
 };